requirements.txt
streamlit==1.32.0
mcp==1.8.0
tavily-python==0.7.2
nest-asyncio==1.6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
langchain==0.2.7
reportlab==4.4.2
uvicorn==0.30.1
starlette==0.45.3
streamlit-cookies-manager==0.2.1
requests==2.32.3
firecrawl==2.5.4
groq==0.15.0
httpx_oauth
llama-index==0.12.44 
llama-index-llms-groq==0.3.2
llama-index-embeddings-huggingface==0.5.5